Turn processes module into a package
