+++ concert/tests/integration/test_processes.py
@@ -3,11 +3,11 @@ import scipy
 from scipy.ndimage import fourier
 from concert.tests import assert_almost_equal, TestCase
 from concert.quantities import q
-from concert.processes import focus
 from concert.devices.cameras.dummy import Base as DummyCameraBase, Camera
 from concert.devices.motors.dummy import LinearMotor, RotationMotor
 from concert.devices.shutters.dummy import Shutter
-from concert.processes import acquire_dark, acquire_image_with_beam, determine_rotation_axis
+from concert.processes.common import focus
+from concert.processes.beamline import acquire_dark, acquire_image_with_beam, determine_rotation_axis
 
 
 MIN_POSITION = 0 * q.mm
